# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/embed/login": {
      "get": {
        "operationId": "EmbedController_page[0]_v1",
        "parameters": [
          {
            "name": "prompt",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "login",
                "auto"
              ],
              "type": "string"
            }
          },
          {
            "name": "display",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "embed",
                "page"
              ],
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "required": false,
            "in": "query",
            "description": "Selects a registered callback",
            "schema": {}
          },
          {
            "name": "origin",
            "required": false,
            "in": "query",
            "description": "Parent origin (exact registered value)",
            "schema": {}
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {}
          },
          {
            "name": "transaction_id",
            "required": true,
            "in": "query",
            "schema": {}
          },
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Core-controlled login page (iframe, or top-level with display=page)",
        "tags": [
          "embed"
        ]
      },
      "post": {
        "operationId": "EmbedController_authenticate[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "description": "Token from the rendered login page",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Authenticate and return the complete RS256 compact assertion for postMessage/form_post delivery",
        "tags": [
          "embed"
        ]
      }
    },
    "/embed/login": {
      "get": {
        "operationId": "EmbedController_page[1]",
        "parameters": [
          {
            "name": "prompt",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "login",
                "auto"
              ],
              "type": "string"
            }
          },
          {
            "name": "display",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "embed",
                "page"
              ],
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "required": false,
            "in": "query",
            "description": "Selects a registered callback",
            "schema": {}
          },
          {
            "name": "origin",
            "required": false,
            "in": "query",
            "description": "Parent origin (exact registered value)",
            "schema": {}
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {}
          },
          {
            "name": "transaction_id",
            "required": true,
            "in": "query",
            "schema": {}
          },
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Core-controlled login page (iframe, or top-level with display=page)",
        "tags": [
          "embed"
        ]
      },
      "post": {
        "operationId": "EmbedController_authenticate[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "description": "Token from the rendered login page",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Authenticate and return the complete RS256 compact assertion for postMessage/form_post delivery",
        "tags": [
          "embed"
        ]
      }
    },
    "/v1/embed/continue": {
      "post": {
        "operationId": "EmbedController_continueSession[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRefDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "SSO: issue an assertion for this client from the existing federation session",
        "tags": [
          "embed"
        ]
      }
    },
    "/embed/continue": {
      "post": {
        "operationId": "EmbedController_continueSession[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRefDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "SSO: issue an assertion for this client from the existing federation session",
        "tags": [
          "embed"
        ]
      }
    },
    "/v1/embed/logout": {
      "post": {
        "operationId": "EmbedController_logout[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "End the central federation session in this browser (\"use a different account\"); no application fan-out",
        "tags": [
          "embed"
        ]
      }
    },
    "/embed/logout": {
      "post": {
        "operationId": "EmbedController_logout[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "End the central federation session in this browser (\"use a different account\"); no application fan-out",
        "tags": [
          "embed"
        ]
      }
    },
...
    "/v1/login": {
      "post": {
        "operationId": "CoreLoginController_signIn[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "POST /login - credentials; returns its_id, name, token, requires_scope_selection, assignments",
        "tags": [
          "core-login"
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "CoreLoginController_signIn[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "POST /login - credentials; returns its_id, name, token, requires_scope_selection, assignments",
        "tags": [
          "core-login"
        ]
      }
    },
...
  },
  "info": {
    "title": "Miqaat Identity Federation Service",
    "description": "Central embedded login, federation sessions, RS256 assertions (JWKS), SSO and federation logout. Authentication only - authorization decisions live in the Core Identity Authorization Service.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EmbedLoginDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "example": "txn-123"
          },
          "client_id": {
            "type": "string",
            "example": "rms-web-prod"
          },
          "identity_type": {
            "type": "string",
            "enum": [
              "ITS",
              "NON_ITS"
            ],
            "default": "ITS"
          },
          "its_id": {
            "type": "string",
            "example": "ITS12345",
            "description": "ITS ID (identity_type ITS)"
          },
          "identifier": {
            "type": "string",
            "example": "member@example.com",
            "description": "Username or email (identity_type NON_ITS)"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "transaction_id",
          "client_id",
          "password"
        ]
      },
      "TransactionRefDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "client_id"
        ]
      },
      "EmbedLogoutDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string"
          }
        },
        "required": [
          "transaction_id"
        ]
      },
      "PortalLoginDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string"
          },
          "identity_type": {
            "type": "string",
            "enum": [
              "ITS",
              "NON_ITS"
            ],
            "default": "ITS"
          },
          "its_id": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "transaction_id",
          "password"
        ]
      },
...
    }
  }
}